Restructure of WindowManager and window size change funcionality
As you can only create one window, all attributes should be static, so
we changed the definition of WindowManaged, changing it from be a class
to be a namespace. Private methods now are static methods inside
system.cpp, and attributes are also static variables in this file.

Now, when you update the window size, it is set to all shaders via
uniform variable 'window_ratio'. All shaders defined should have this
uniform variable, although they don't use it.