Added support for texture loading
Now you can load textures from BMP images using Texture class.
These images can be RGB or RGBA. They should be squared, but
they could be rectangled id you have modern graphics card.

loadSquare function now accepts a new parameter, which is a
path to a BMP image. If this path is empty (0 chars), texture
will not be loaded. Otherwise, it will be loaded and used to
draw the square.