Added a basic AI which plays Snake
If you are bored, and you want to watch a gameplay, you can run this
game which plays for you!. It implements a basic algorithm (Breadth-first
search). The efficiency of this algorithm is O(n^2·3^n). It's exponential
because of the search, and cuadratic because we copy in each iteration
a new board, which has n·n elements.